requirements.txt
Flask==2.0.1
Pillow==8.0.1
python-docx==0.8.11
pandas==1.1.3